# docs/swagger.yaml
basePath: /
definitions:
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  dto.LoginResponse:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: string
      tokenType:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.SignUpUserRequest:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      password:
        type: string
      topicsOfInterest:
        items:
          $ref: '#/definitions/model.TopicOfInterest'
        type: array
      username:
        type: string
    type: object
  dto.SignUpUserResponse:
    properties:
      firstname:
        type: string
      lastname:
        type: string
      username:
        type: string
    type: object
...
  dto.StatisticsResponse:
    properties:
      timeSpentOnTeams:
        items:
          $ref: '#/definitions/model.TimeSpentOnTeam'
        type: array
      totalTimeSpentOnApp:
        example: 7200000
        type: integer
      userId:
        type: string
    type: object
...
  dto.UpdateStatisticsRequest:
    properties:
      teamId:
        example: team123
        type: string
      timeSpentOnApp:
        example: 1800000
        type: integer
      timeSpentOnTeam:
        example: 900000
        type: integer
    type: object
  dto.UserPasswordRequestDTO:
    properties:
      id:
        type: string
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      statistics:
        $ref: '#/definitions/model.Statistics'
      teams:
        items:
          type: string
        type: array
      topicsOfInterest:
        items:
          $ref: '#/definitions/model.TopicOfInterest'
        type: array
      username:
        type: string
    type: object
...
  dto.UserUpdateRequestDTO:
    properties:
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      topicsOfInterest:
        items:
          $ref: '#/definitions/model.TopicOfInterest'
        type: array
      username:
        type: string
    type: object
  dto.UserUpdateResponseDTO:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      statistics:
        $ref: '#/definitions/model.Statistics'
      teams:
        items:
          type: string
        type: array
      topicsOfInterest:
        items:
          $ref: '#/definitions/model.TopicOfInterest'
        type: array
      username:
        type: string
    type: object
...
  entity.User:
    properties:
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      password:
        type: string
      statistics:
        $ref: '#/definitions/model.Statistics'
      teams:
        items:
          type: string
        type: array
      topicsOfInterest:
        items:
          $ref: '#/definitions/model.TopicOfInterest'
        type: array
      username:
        type: string
    type: object
...
  model.Statistics:
    properties:
      timeSpentOnTeams:
        items:
          $ref: '#/definitions/model.TimeSpentOnTeam'
        type: array
      totalTimeSpentOnApp:
        example: 7200000
        type: integer
    type: object
  model.TimeSpentOnTeam:
    properties:
      duration:
        example: 3600000
        type: integer
      teamId:
        type: string
    type: object
  model.TopicOfInterest:
    enum:
    - Mathematics
    - Science
    - ComputerScience
    - Programming
    - Art
    - Music
    - Photography
    - Language
    type: string
    x-enum-varnames:
    - Mathematics
    - Science
    - ComputerScience
    - Programming
    - Art
    - Music
    - Photography
    - Language
info:
  contact: {}
  title: StudyWithMe API
  version: "1.0"
paths:
...
  /users:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: The user's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Delete a user
      tags:
      - users
    get:
      consumes:
      - application/json
      parameters:
      - description: The user's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a user by ID
      tags:
      - users
    patch:
      consumes:
      - application/json
      parameters:
      - description: The user's ID
        in: path
        name: id
        required: true
        type: string
      - description: The user profile update (all fields optional)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserUpdateRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserUpdateResponseDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user profile (selective fields)
      tags:
      - users
  /users/{id}/friends:
    get:
      description: Get list of friends for a user (accepted requests)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get friends for a user
      tags:
      - users
  /users/{id}/mutual/{otherId}:
    get:
      description: Get list of mutual friends between userA and userB
      parameters:
      - description: User A ID
        in: path
        name: id
        required: true
        type: string
      - description: User B ID
        in: path
        name: otherId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.User'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get mutual friends between two users
      tags:
      - users
  /users/{id}/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: The user's ID
        in: path
        name: id
        required: true
        type: string
      - description: The password update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserPasswordRequestDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user password
      tags:
      - users
  /users/{id}/statistics:
    get:
      parameters:
      - description: The user's ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatisticsResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get a user's statistics
      tags:
      - users
    put:
      consumes:
      - application/json
      parameters:
      - description: The user's ID
        in: path
        name: id
        required: true
        type: string
      - description: The statistics update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStatisticsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatisticsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user statistics
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Accepts either `email` or `username` along with `password`. Returns
        an access token and the full user (without password).
      parameters:
      - description: The login request (email or username + password)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user by email or username and return JWT
      tags:
      - users
  /users/signup:
    post:
      consumes:
      - application/json
      parameters:
      - description: The sign-up request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SignUpUserResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
...
securityDefinitions:
  Bearer:
    description: 'Bearer JWT token. Example: "Bearer YOUR_TOKEN"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"